requirements.txt
fastapi
uvicorn
openai
chromadb
python-dotenv
requests
httpx
woocommerce
sentence-transformers
